+++ internal/envoy/envoy_linux.go
@@ -5,6 +5,7 @@ package envoy
 import (
 	"context"
 	"io/ioutil"
+	"os/exec"
 	"strconv"
 	"sync"
 	"syscall"
@@ -26,6 +27,16 @@ var restartEpoch struct {
 var sysProcAttr = &syscall.SysProcAttr{
 	Setpgid:   true,
 	Pdeathsig: syscall.SIGTERM,
+}
+
+// Command creates an exec.Cmd using the embedded envoy binary.
+func Command(ctx context.Context, arg ...string) (*exec.Cmd, error) {
+	fullEnvoyPath, err := setup()
+	if err != nil {
+		return nil, err
+	}
+
+	return exec.CommandContext(ctx, fullEnvoyPath, arg...), nil
 }
 
 func (srv *Server) runProcessCollector(ctx context.Context) {
